Update gradle project setup and libraries.
Use the new gradle plugin dsl and update the android gradle plugin to version 8.2.
Change the workflow to use java 17, which is required by the android gradle plugin.
The android gradle plugin now provides the sources and javadoc jars, which replaces the custom task setup.
To update the publishing setup, the related code has been inlined with old references removed.
Increase the compile sdk to 34 and the kotlin version to 1.9.22.